// frontend/src/shared/schemas/schemas.ts
// Auto-generated by @cerios/openapi-to-zod
// API: Busines_Club API
// Version: 1.0
// Do not edit this file manually

import { z } from "zod";

// Schemas and Types
export const objectIdSchema = z.strictObject({});
export type ObjectId = z.infer<typeof objectIdSchema>;

export const peopleDtoSchema = z.array(
  z.strictObject({
    first_name: z.string(),
    last_name: z.string(),
    age: z.number(),
    interests: z.array(z.string()),
    description: z.string(),
    tg_username: z.string(),
    type_activity: z.string(),
    photo_url: z.url(),
    isLiked: z.boolean().optional(),
    _id: z.string(),
  }),
);
export type PeopleDto = z.infer<typeof peopleDtoSchema>;

export const eventDtoSchema = z.array(
  z.strictObject({
    start_time: z.string().regex(/\/^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$\//),
    end_time: z.string().regex(/\/^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$\//),
    speaker_name: z.string(),
    company_name: z.string(),
    title: z.string(),
    place: z.string(),
    isLiked: z.boolean().optional(),
    photo_url: z.url().optional(),
    _id: z.string(),
  }),
);
export type EventDto = z.infer<typeof eventDtoSchema>;